requirements.txt
Django
pyrebase4
django-cors-headers
djangorestframework
django-summernote
Pillow
mysqlclient
jmespath
itsdangerous
mysql-connector-python